Pipeline TEST

Related work items: #232, #235, #236, #238, #241, #267, #268, #269, #270, #272, #273, #274, #275, #280, #307, #309, #310, #311, #316, #317, #318, #319, #320, #321, #322, #325, #331, #337, #338, #349, #351, #355, #359, #363

+++ DijiWalk.Entities/StepValidation.cs
@@ -38,7 +38,7 @@ namespace DijiWalk.Entities
         /// <summary>
         /// Obtient ou définit le score
         /// </summary>
-        public float Score { get; set; }
+        public double Score { get; set; }
 
         /// <summary>
         /// Obtient ou définit l'étape
